requirements.txt
torch
torchvision
numpy
pandas
tqdm
matplotlib
Pillow

